Add another missing cast

+++ gdal/ogr/ogrsf_frmts/flatgeobuf/packedrtree.cpp
@@ -361,7 +361,7 @@ void PackedRTree::streamWrite(const std::function<void(uint8_t *, size_t)> &writ
         CPL_LSBPTR64(&_nodeItems[i].offset);
     }
 #endif
-    writeData(reinterpret_cast<uint8_t *>(_nodeItems), _numNodes * sizeof(NodeItem));
+    writeData(reinterpret_cast<uint8_t *>(_nodeItems), static_cast<size_t>(_numNodes * sizeof(NodeItem)));
 }
 
 NodeItem PackedRTree::getExtent() const { return _extent; }
